refactor: 단일 사용 유틸리티/컴포넌트를 라우트별 private folder로 이동
