+++ app/(home)/_hooks/index.ts
@@ -1,3 +1,5 @@
 export { default as useCategoryQueryParam } from './useCategoryQueryParam';
+export { default as useInfiniteScroll } from './useInfiniteScroll';
 export { default as useInfiniteScrollPostList } from './useInfiniteScrollPostList';
+export { default as useIntersectionObserver } from './useIntersectionObserver';
 export { default as useTiltEffect } from './useTiltEffect';
