+++ app/(home)/_hooks/useIntersectionObserver.ts
