+++ app/(home)/_utils/getCategories.ts
