+++ app/(home)/_utils/index.ts
@@ -1,0 +1,2 @@
+export { default as getBlurImage } from './getBlurImage';
+export { default as getCategories } from './getCategories';
