+++ app/_components/layout/index.tsx
@@ -1,7 +1,7 @@
 import { PropsWithChildren } from 'react';
 
-import Footer from '@/components/layout/Footer';
-import Header from '@/components/layout/Header';
+import Footer from './Footer';
+import Header from './Header';
 
 const Layout = ({ children }: PropsWithChildren) => {
   return (
