--- src/hooks/index.ts
@@ -1,2 +1,0 @@
-export { default as useInfiniteScroll } from './useInfiniteScroll';
-export { default as useIntersectionObserver } from './useIntersectionObserver';
