+++ src/utils/index.ts
@@ -1,9 +1,4 @@
-export { cachedFetchIdBySlug } from './fetchIdBySlug';
-export { cachedFetchNotionPageProperties } from './fetchNotionPageProperties';
 export { cachedFetchNotionPostsMeta } from './fetchNotionPostsMeta';
 export { cachedFetchNotionProfileUrl } from './fetchNotionProfileUrl';
-export { default as getBlurImage } from './getBlurImage';
-export { default as getCategories } from './getCategories';
 export { default as getPostsMeta } from './getPostsMeta';
-export { default as getSlugs } from './getSlugs';
 export { default as notionClient } from './notionClient';
